Fix preventing validation hooks when save! is used
Added a spec causing the valid? call to be run when save! is used,
to work around it, I've added an extra check to the save_self
chainable in dm-validations.rb that checks for the execute_hooks
parameter, circumventing a call to valid?

Note: I don't think this is an actual fix, because I think it has
to do with the way the validation_context_stack is used,
especially the check for emptyness in the save_self chainable.
Apparently it is possible to get in a state in which the resource
is dirty and the stack is not empty, resulting in a call to valid?